requirements.txt
pandas
faiss-cpu
openai
python-dotenv